@@ -23,10 +23,10 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^20.1.0",
+    "@angular/common": "^20.2.4",
     "@angular/compiler": "^20.1.0",
-    "@angular/core": "^20.1.0",
-    "@angular/forms": "^20.1.0",
+    "@angular/core": "^20.2.4",
+    "@angular/forms": "^20.2.2",
     "@angular/platform-browser": "^20.2.2",
     "@angular/platform-server": "^20.1.0",
     "@angular/router": "^20.1.0",
