@@ -1,11 +1,11 @@
 import { Component } from '@angular/core';
-
+import { RouterOutlet } from '@angular/router';
 
 @Component({
   selector: 'app-header-corretor',
+   standalone: true,
+  imports: [RouterOutlet],
   templateUrl: './header-corretor.component.html',
   styleUrls: ['./header-corretor.component.scss'],
 })
-export class HeaderAdminComponent {
-
-}
+export class HeaderCorretorComponent {}   // corrigido
