@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { AuthService } from 'src/app/services/auth.service';
+import { AuthService } from '../../services/auth.service';
 
 @Component({
   selector: 'app-interna',
@@ -8,18 +8,18 @@ import { AuthService } from 'src/app/services/auth.service';
   styleUrls: ['./interna.component.scss']
 })
 export class InternaComponent implements OnInit {
-  cliente: any = {};
+  usuario: any = {};
 
   constructor(private authService: AuthService, private router: Router) {}
 
   ngOnInit() {
-    this.carregarCliente();
+    this.carregarUsuario();
   }
 
-  carregarCliente() {
-    this.cliente = this.authService.clienteAtual();
+  carregarUsuario() {
+    this.usuario = this.authService.usuarioAtual();
     
-    if (!this.cliente || Object.keys(this.cliente).length === 0) {
+    if (!this.usuario || Object.keys(this.usuario).length === 0) {
       this.router.navigate(['/login']);
     }
   }
